Fix react loadable example (#418)
* react-loadable example: fix bundle loading order

in the react-loadable example the bundles are currently not loaded in
the correct order: the "chunks" are loaded before the main bundle is
loaded. This does not work because the webpack bundle functions required
to load the chunks are not yet loaded, leading to this error:
"0.51b7694d.chunk.js:1 Uncaught ReferenceError: webpackJsonp is not
defined"

This commit changes server.js to follow the react-loadable docs more
closely. It removes the defer attribute from the main bundle, so that it
is loaded before the chunks.
Also uses Loadable.preloadReady() to defer hydration until all chunks
are loaded, as described in the react-loadable docs.

* react-loadable example: fix chunk path in dev

when running the dev-server, the preloaded chunks for react-loadable
must be loaded from the dev-server running at port 3001(by default).
Before this commit they were loaded with relative path - this worked in
production, but in dev mode the index.html was served instead, resulting
in "Uncaught SyntaxError: Unexpected token <" errors when parsing the
chunks.